change tag order
